+++ openai-java-example/src/main/java/com/openai/example/StructuredOutputsAsyncExample.java
@@ -1,0 +1,80 @@
+package com.openai.example;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.openai.client.OpenAIClientAsync;
+import com.openai.client.okhttp.OpenAIOkHttpClientAsync;
+import com.openai.models.ChatModel;
+import com.openai.models.chat.completions.ChatCompletionCreateParams;
+import com.openai.models.chat.completions.StructuredChatCompletionCreateParams;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
+
+public final class StructuredOutputsAsyncExample {
+
+    public static class Person {
+        @JsonPropertyDescription("The first name and surname of the person.")
+        public String name;
+
+        public int birthYear;
+
+        @JsonPropertyDescription("The year the person died, or 'present' if the person is living.")
+        public String deathYear;
+
+        @Override
+        public String toString() {
+            return name + " (" + birthYear + '-' + deathYear + ')';
+        }
+    }
+
+    public static class Book {
+        public String title;
+
+        public Person author;
+
+        @JsonPropertyDescription("The year in which the book was first published.")
+        @Schema(minimum = "1500")
+        public int publicationYear;
+
+        public String genre;
+
+        @JsonIgnore
+        public String isbn;
+
+        @Override
+        public String toString() {
+            return '"' + title + "\" (" + publicationYear + ") [" + genre + "] by " + author;
+        }
+    }
+
+    public static class BookList {
+        @ArraySchema(maxItems = 100)
+        public List<Book> books;
+    }
+
+    private StructuredOutputsAsyncExample() {}
+
+    public static void main(String[] args) {
+        // Configure using one of:
+        // - The `OPENAI_API_KEY` environment variable
+        // - The `OPENAI_BASE_URL` and `AZURE_OPENAI_KEY` environment variables
+        OpenAIClientAsync client = OpenAIOkHttpClientAsync.fromEnv();
+
+        StructuredChatCompletionCreateParams<BookList> createParams = ChatCompletionCreateParams.builder()
+                .model(ChatModel.GPT_4O_MINI)
+                .maxCompletionTokens(2048)
+                .responseFormat(BookList.class)
+                .addUserMessage("List some famous late twentieth century novels.")
+                .build();
+
+        client.chat()
+                .completions()
+                .create(createParams)
+                .thenAccept(completion -> completion.choices().stream()
+                        .flatMap(choice -> choice.message().content().stream())
+                        .flatMap(bookList -> bookList.books.stream())
+                        .forEach(book -> System.out.println(" - " + book)))
+                .join();
+    }
+}